+++ lib/assets/domain/domain.rb
@@ -4,52 +4,28 @@ include Punch
 # Design your domain here
 # look through sample.rb for details
 def build_domain
-  DSL::Builder.build('Domain', 'Descriptioin') do
+  DSL::Builder.build do
+    # sentry "User Name",  'must be String[3, 100]'
+    # sentry :email,  'must be valid Email addresss'
+    # sentry :secret, 'at least 8 symbols with digits'
 
-    sentry :string, block: 'v.is_a?(Strinig)'
-    sentry :email,  'must be valid email address'
-    sentry :secret, 'at least 8 symbols with digits'
+    # entity :user do
+    #   param :login, :sentry => :email
+    #   param :secret, :sentry => :secret
+    # end
 
-    entity :user do
-      param :login, :sentry => :email
-      param :secret, :sentry => :secret
-    end
+    # core service
+    # service "Login" do
+    #   param :email,  :sentry => :email
+    #   param :secret, :sentry => :secret
+    # end
 
-    entity :credentials do
-      param :login, :sentry => :email
-      param :secret, :sentry => :secret
-    end
-
-    actor :user do
-      service :signup do
-        param :login, :sentry => :email
-        param :secret, :sentry => :secret
-      end
-
-      service :signin do
-        param :login, :sentry => :email
-        param :secret, :sentry => :secret
-      end
-
-      service :forget do
-        param :login, :sentry => :email
-        param :secret, :sentry => :secret
-      end
-    end
-
-    actor :admin do
-      service :query_users do
-        param :page_number, :sentry => :page_number
-        param :page_size, :sentry => :page_size
-      end
-
-      service :lock_user do
-        param :login, :sentry => :email
-      end
-
-      service :unlock_user do
-        param :login, :sentry => :email
-      end
-    end
+    # actors services
+    # actor :user do
+    #   service "Sing Up" do
+    #     param :name,  :sentry => :user_name
+    #     param :email, :sentry => :email
+    #   end
+    # end
   end
 end