+++ test/punch/services/test_punch_sentry.rb
@@ -49,7 +49,7 @@ end
 
 class PunchSentry
   public_class_method :new
-  public :storage, :punch, :render, :declared?, :declared, :sentries_source
+  public :storage, :punch, :render, :declared?, :declared
   attr_accessor :source, :test
 end