+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 
+## [0.5.1] - 2023-01-06
+
+- added test for `punch/basics`
+- added `PunchDomain` service
+- changed `domain/dogen.rb` for using `PunchDomain`
+- optimized samples
+- optimized README
+
 ## [0.5.0] - 2023-01-06
 
 - source code completely redesigned